refactor: now each component is on a specific file
