+++ src/components/ProductCard.tsx
@@ -4,7 +4,7 @@
  * File Created: Monday, 31st August 2020 3:33:49 pm
  * Author: Esperanza Horn (esperanza@inventures.cl)
  * -----
- * Last Modified: Thursday, 10th September 2020 5:51:09 pm
+ * Last Modified: Friday, 11th September 2020 10:54:14 am
  * Modified By: Esperanza Horn (esperanza@inventures.cl)
  * -----
  * Copyright 2020 - 2020 Incrementa Ventures SpA. ALL RIGHTS RESERVED
@@ -15,22 +15,18 @@
 
 import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
-import ChevronRightRoundedIcon from '@material-ui/icons/ChevronRightRounded';
 import {
   Card,
   CardActionArea,
-  Grid,
-  IconButton,
   CardMedia,
   CardContent,
   Typography,
   Chip,
-  Box,
 } from '@material-ui/core';
 import { CurrencyFormatter } from '../formatters';
 import clsx from 'clsx';
 
-type ProductBPropTypes = {
+export type ProductPropTypes = {
   imageUrl?: string;
   title: string;
   subtitle?: string;
@@ -41,103 +37,12 @@ type ProductBPropTypes = {
   price: number;
   onClickCard: () => void;
 };
-type ProductListHeader = {
-  title?: string;
-  onClickCarousel?: () => void;
-};
-type ProductList = {
-  products: ProductBPropTypes[];
-  gridBreakpoints: GridBreakpoints;
-};
-type GridBreakpoints = {
-  xs:
-    | boolean
-    | 'auto'
-    | 1
-    | 2
-    | 3
-    | 4
-    | 5
-    | 6
-    | 7
-    | 8
-    | 9
-    | 10
-    | 11
-    | 12
-    | undefined;
-  sm:
-    | boolean
-    | 'auto'
-    | 1
-    | 2
-    | 3
-    | 4
-    | 5
-    | 6
-    | 7
-    | 8
-    | 9
-    | 10
-    | 11
-    | 12
-    | undefined;
-  md:
-    | boolean
-    | 'auto'
-    | 1
-    | 2
-    | 3
-    | 4
-    | 5
-    | 6
-    | 7
-    | 8
-    | 9
-    | 10
-    | 11
-    | 12
-    | undefined;
-  lg:
-    | boolean
-    | 'auto'
-    | 1
-    | 2
-    | 3
-    | 4
-    | 5
-    | 6
-    | 7
-    | 8
-    | 9
-    | 10
-    | 11
-    | 12
-    | undefined;
-  xl:
-    | boolean
-    | 'auto'
-    | 1
-    | 2
-    | 3
-    | 4
-    | 5
-    | 6
-    | 7
-    | 8
-    | 9
-    | 10
-    | 11
-    | 12
-    | undefined;
-};
 
 const useStyles = makeStyles({
   root: {
     textAlign: 'center',
     padding: '8px',
   },
-
   title: {
     width: '100%',
     margin: '4px 0px 8px 0px',
@@ -157,24 +62,15 @@ const useStyles = makeStyles({
     marginTop: '-14px',
     maxWidth: '100%',
   },
-  carouselHeader: {
-    paddingBottom: '8px',
-    display: 'flex',
-    justifyContent: 'space-between',
-    alignItems: 'flex-end',
-  },
   disabledTag: {
     backgroundColor: '#FFFFFF',
   },
   price: {
     marginTop: '8px',
   },
-  viewMore: {
-    padding: '0px',
-  },
 });
 
-export function ProductCard(props: ProductBPropTypes) {
+export function ProductCard(props: ProductPropTypes) {
   const {
     imageUrl,
     title,
@@ -243,59 +139,5 @@ export function ProductCard(props: ProductBPropTypes) {
         </CardContent>
       </CardActionArea>
     </Card>
-  );
-}
-
-export function ProductList(props: ProductList) {
-  const {
-    products,
-    gridBreakpoints: { xs = 6, sm = 4, md = 3, lg = 2, xl = 1 } = {},
-  } = props;
-  return (
-    <Grid container spacing={1}>
-      {products.map((cardInfo: ProductBPropTypes, index: number) => (
-        <Grid key={index} item xs={xs} sm={sm} md={md} lg={lg} xl={xl}>
-          <ProductCard
-            imageUrl={cardInfo.imageUrl}
-            title={cardInfo.title}
-            subtitle={cardInfo.subtitle}
-            details={cardInfo.details}
-            description={cardInfo.description}
-            price={cardInfo.price}
-            tagText={cardInfo.tagText}
-            tagIcon={cardInfo.tagIcon}
-            onClickCard={cardInfo.onClickCard}
-          />
-        </Grid>
-      ))}
-    </Grid>
-  );
-}
-
-export function ProductListHeader(props: ProductListHeader) {
-  const { title, onClickCarousel } = props;
-  const classes = useStyles();
-
-  return (
-    <Box>
-      {title && (
-        <Box className={classes.carouselHeader}>
-          <Typography variant="h6" color="textPrimary">
-            {title}
-          </Typography>
-
-          {onClickCarousel && (
-            <IconButton
-              color="primary"
-              aria-label="ver más resultados"
-              onClick={onClickCarousel}
-              className={classes.viewMore}
-            >
-              <ChevronRightRoundedIcon fontSize="large" />
-            </IconButton>
-          )}
-        </Box>
-      )}
-    </Box>
   );
 }
