+++ src/components/index.ts
@@ -4,7 +4,7 @@
  * File Created: Friday, 31st July 2020 3:20:09 pm
  * Author: Gabriel Ulloa (gabriel@inventures.cl)
  * -----
- * Last Modified: Thursday, 3rd September 2020 5:54:54 pm
+ * Last Modified: Friday, 11th September 2020 11:06:00 am
  * Modified By: Esperanza Horn (esperanza@inventures.cl)
  * -----
  * Copyright 2019 - 2020 Incrementa Ventures SpA. ALL RIGHTS RESERVED
@@ -16,3 +16,9 @@ export * from './input';
 export * from './ProductCard';
 export * from './searchBar';
 export * from './CartProduct';
+export * from './ProductDetails';
+export * from './ProductList';
+export * from './ProductListHeader';
+
+
+
